feat: add Player struct
